buxfix:  error:0308010C:digital envelope routines::unsupported
changes script start

The error can also occur if you build the application using docker build since it pulls the latest version of Node.JS by default.

This error is because node.js 17 uses OpenSSL3, which has changed code for initialization context of md family (including md4), and this is a breaking change.